Fixed #4; add option to run `git fetch`.
Added new setting:
`flexprompt.settings.git_fetch_interval` can be set to a number, to
cause the prompt to also run `git fetch` periodically.

When set, the `{git}` module will automatically run `git fetch` when the
current directory is in a repo if either of the following are true:
- It's the first time the repo has been
visited in this session.
- It's been more than the specified number of minutes since the prompt
ran `git fetch` for the repo.