Colour a git branch in magenta if not published
The name if the branch is colored in magenta if no upstream is configured
for the branch.

This code adds --branch to git staus, as in

git --no-optional-locks status --porcelain --branch 2>nul

The first line of git status is then processed to get summary information
of the branch.

If the branch produces a summary line (e.g., valid git repo), returns the name
of the branch but not the upstream information, then the branch is considered
unpublished and it is colored in magenta (git_colors.staged).