Improve powerline font support (breaking change).
Now the symbols table can specify which symbols require a powerline
font, and can fall back to other symbols if a powerline font is not
available.

BREAKING CHANGE:  The above change has a side effect of making existing
wizard configs stop showing the branch symbol until the wizard is run to
update the config.