Add `\003` and `\004` color codes.
Prompt segment text and separators can use these codes, as well as the
`\001` and `\002` codes.

The new codes make it easy to apply the frame color or separator color
within prompt segment text or a separator.

For example, the following makes a separator that is a frame-colored
minus sign:
`flexprompt.settings.separators = "\003-"`