Fix `{cwd:shorten}` with uppercase letters.
When the first letter in a word was uppercase, the a mistake in the
pattern string caused it to fail to match.

When `%HOME%` env var contains `%` or `.` or `[` or `^` or `$` then the
`use_home_tilde` setting failed to replace the home dir with a tilde.

The docs refer to `shorten=rootsmart`, but the code recognized
`shorten=smartroot`.  Now it accepts both spellings.

Plus a few minor code cleaniness adjustments.