Update vpn and anyconnect modules to use `prompt_info`.
Flexprompt has documentation about how to easily make async prompt
modules.  The `vpn` module didn't need part of that, so it just used
`flexprompt.promptcoroutine()`.  But that makes it harder to learn from
examples.

So now all the built-in modules use the simpler (and documented)
approach of calling `flexprompt.prompt_info()`.