Update README sample.
The README didn't mention the "best practice" initialization pattern.
The flexprompt_autoconfig.lua script uses it, so following the
recommendation to create a flexprompt_config.lua by copying from
flexprompt_autoconfig.lua works well.  But following the
documentation to create a flexprompt_config from scratch would miss the
initialization pattern and could result in errors if the scripts are in
different directories.