Fix #63; off by one when getting battery symbol.
When using {battery:levelicon} the battery level icon was one off from
what it was meant to be (full battery showed a 90% icon, and 10% showed
a 0% icon).

That caused the index to be 0 when the battery level was below 10%,
which caused a script error since Lua array indexing is one-based.