Add support for Smart Charging.
The battery prompt module can detect Smart Charging and show a symbol
for it next to the battery level.  By default the symbol is a heart.

Smart Charging is an operating system feature where it adjusts the
battery charging to help prolong the service lifetime of the battery.
When plugged in for an extended period of time the laptop battery may
deplete below 100% without necessarily immediately charging it to 100%.