Config wizard allows choosing new blur styles.
When a terminal supports 24 bit color and the "Blurred" style is chosen for
Heads or Tails (e.g. in Classic or Rainbow styles), then the config wizard
shows an additional prompt to ask which blur effect to use (block graphics,
fade to black, or fade to white).  If either of the "fade" effects are chosen,
then the config wizard also asks which blur effect width to use (from 1 to 5
characters wide).